requirements.txt
amqp==5.0.9
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==21.12b0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==4.0
django-heroku==0.0.0
django-phone-field==1.8.1
gunicorn==20.1.0
kombu==5.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
